requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
easydict==1.13
fonttools==4.53.0
ftfy==6.2.0
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
prefetch_generator==1.0.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scipy==1.13.1
six==1.16.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.19.2
